Fixes after testing

+++ pkg/sources/git/git.go
@@ -299,7 +299,11 @@ func (s *Source) scanRepo(ctx context.Context, repoURI string, reporter sources.
 
 	err := func() error {
 		path, repo, err := cloneFunc()
-		defer os.RemoveAll(path)
+		// remove the directory only if it was created as a temporary path, or if it is a clone path and --no-cleanup is not set.
+		if strings.HasPrefix(path, filepath.Join(os.TempDir(), "trufflehog")) || (!s.conn.GetNoCleanup() && s.conn.GetClonePath() != "") {
+			defer os.RemoveAll(path)
+		}
+
 		if err != nil {
 			return err
 		}
@@ -349,7 +353,7 @@ func (s *Source) scanDir(ctx context.Context, gitDir string, reporter sources.Ch
 
 	err = func() error {
 		// remove the directory only if it was created as a temporary path, or if it is a clone path and --no-cleanup is not set.
-		if strings.HasPrefix(gitDir, filepath.Join(os.TempDir(), "trufflehog")) || (!s.conn.NoCleanup && s.conn.GetClonePath() != "") {
+		if strings.HasPrefix(gitDir, filepath.Join(os.TempDir(), "trufflehog")) || (!s.conn.GetNoCleanup() && s.conn.GetClonePath() != "") {
 			defer os.RemoveAll(gitDir)
 		}
 
@@ -411,7 +415,7 @@ func CloneRepo(ctx context.Context, userInfo *url.Userinfo, gitURL string, clone
 
 	// If --clone-path is set, create a subdirectory <clonePath>/<repo-name> with permissions 0755.
 	if clonePath != "" {
-		path = filepath.Join(clonePath, filepath.Base(gitURL))
+		path = filepath.Join(clonePath, strings.TrimSuffix(filepath.Base(gitURL), ".git"))
 		if err = os.MkdirAll(path, 0755); err != nil {
 			return "", nil, fmt.Errorf("failed to create clone path %s: %w", clonePath, err)
 		}
